+++ resources/views/components/breadcrumbs.blade.php
@@ -8,22 +8,22 @@
 
 @if (!empty($path))
   <div
-    class="{{ $classes ?? '' }} wbtr:flex wbtr:flex-row wbtr:flex-wrap wbtr:items-center wbtr:gap-2"
+    class="{{ $classes ?? '' }} w-breadcrumbs wbtr:flex wbtr:flex-row wbtr:flex-wrap wbtr:items-center wbtr:gap-2"
   >
     @foreach ($path as $item)
       @if ($item[1] !== '')
         <a
           href="{{ $item[1] }}"
-          class="wbtr:text-caption {{ $loop->last ? 'wbtr:text-black' : 'wbtr:text-gray-500' }} wbtr:hover:underline"
+          class="w-breadcrumbs__item wbtr:text-caption {{ $loop->last ? 'wbtr:text-black' : 'wbtr:text-gray-500' }} wbtr:hover:underline"
         >
         @else
-          <div class="wbtr:text-caption {{ $loop->last ? 'wbtr:text-black' : 'wbtr:text-gray-500' }}">
+          <div class="w-breadcrumbs__item wbtr:text-caption {{ $loop->last ? 'wbtr:text-black' : 'wbtr:text-gray-500' }}">
       @endif
       {!! $item[0] !!}
       </{{ $item[1] !== '' ? 'a' : 'div' }}>
 
       @if (!$loop->last)
-        @svg('images.svg.chevron-right')
+        @svg('images.svg.chevron-right', 'w-breadcrumbs__icon')
       @endif
     @endforeach
   </div>