+++ resources/views/components/tab-container.blade.php
@@ -17,16 +17,16 @@
   id="{{ $id }}"
   x-show="activeTab === '{{ $id }}'"
   x-cloak
-  class="tab-container tab-panel"
+  class="w-tab-container tab-container tab-panel"
   x-transition:enter-start="wbtr:opacity-0 wbtr:scale-100"
   x-transition:enter="wbtr:transition wbtr:ease-in wbtr:duration-500"
   x-transition:enter-end="wbtr:opacity-100 wbtr:scale-100"
 >
   <div
-    class="{{ $classes ?? '' }} tab-container__content wbtr:flex wbtr:w-full wbtr:flex-col wbtr:items-start wbtr:justify-start"
+    class="{{ $classes ?? '' }} w-tab-container__content tab-container__content wbtr:flex wbtr:w-full wbtr:flex-col wbtr:items-start wbtr:justify-start"
   >
     @if (!empty($title) && empty($hide_title))
-      <h2 class="text-h2">{{ $title }}</h2>
+      <h2 class="w-tab-container__title text-h2">{{ $title }}</h2>
     @endif
 
     @if ($tab_content)